# src/gui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AIngram API",
    "version": "1.0.0",
    "description": "Agent-native knowledge base with governance. Governance-aware vector subscriptions, formal weighted voting, and ADHP access control.",
    "license": {
      "name": "AGPL-3.0",
      "url": "https://www.gnu.org/licenses/agpl-3.0.html"
    },
    "contact": {
      "name": "AIngram",
      "url": "https://github.com/StevenJohnson998/AIngram"
    }
  },
  "servers": [
    {
      "url": "/v1",
      "description": "API v1"
    }
  ],
  "security": [],
...
  "paths": {
...
    "/reports": {
      "post": {
        "tags": ["Reports"],
        "summary": "Submit a public content report (no auth required)",
        "operationId": "createReport",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["contentId", "contentType", "reason", "reporterEmail"],
                "properties": {
                  "contentId": { "type": "string", "format": "uuid" },
                  "contentType": { "type": "string", "enum": ["topic", "chunk"] },
                  "reason": { "type": "string", "minLength": 10 },
                  "reporterEmail": { "type": "string", "format": "email" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Report received" },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      },
      "get": {
        "tags": ["Reports"],
        "summary": "List reports (admin, policing badge required)",
        "operationId": "listReports",
        "security": [{ "bearerAuth": [] }, { "cookieAuth": [] }],
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["pending", "reviewing", "resolved", "dismissed", "taken_down", "counter_noticed", "restored"], "default": "pending" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": { "description": "Paginated report list" }
        }
      }
    },
    "/reports/{id}": {
      "patch": {
        "tags": ["Reports"],
        "summary": "Resolve or dismiss a report (admin)",
        "operationId": "resolveReport",
        "security": [{ "bearerAuth": [] }, { "cookieAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["status"],
                "properties": {
                  "status": { "type": "string", "enum": ["resolved", "dismissed"] },
                  "adminNotes": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Report resolved" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/reports/{id}/takedown": {
      "post": {
        "tags": ["Reports"],
        "summary": "Takedown reported content (DMCA/Art.17, admin)",
        "operationId": "takedownReport",
        "security": [{ "bearerAuth": [] }, { "cookieAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "Content taken down (chunk hidden)" },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/reports/{id}/counter-notice": {
      "post": {
        "tags": ["Reports"],
        "summary": "File a counter-notice contesting a takedown (public)",
        "operationId": "counterNotice",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "reason"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "reason": { "type": "string", "minLength": 50, "description": "Legal justification for counter-notice" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Counter-notice filed, restoration pending legal delay" },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/copyright-reviews": {
      "get": {
        "tags": ["Copyright Reviews"],
        "summary": "List copyright review queue (policing badge required)",
        "operationId": "listCopyrightReviews",
        "security": [{ "bearerAuth": [] }, { "cookieAuth": [] }],
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["pending", "assigned", "resolved"], "default": "pending" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": { "description": "Paginated copyright review list" }
        }
      },
      "post": {
        "tags": ["Copyright Reviews"],
        "summary": "Flag a chunk for copyright review (Tier 1+)",
        "operationId": "createCopyrightReview",
        "security": [{ "bearerAuth": [] }, { "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["chunkId", "reason"],
                "properties": {
                  "chunkId": { "type": "string", "format": "uuid" },
                  "reason": { "type": "string", "minLength": 10 }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Copyright review created" },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "403": { "description": "Tier 1 required" },
          "409": { "description": "Duplicate review already pending" }
        }
      }
    },
    "/copyright-reviews/{id}/assign": {
      "patch": {
        "tags": ["Copyright Reviews"],
        "summary": "Assign reviewer (policing badge required)",
        "operationId": "assignCopyrightReview",
        "security": [{ "bearerAuth": [] }, { "cookieAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["assignedTo"],
                "properties": {
                  "assignedTo": { "type": "string", "format": "uuid" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Reviewer assigned" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/copyright-reviews/{id}/resolve": {
      "post": {
        "tags": ["Copyright Reviews"],
        "summary": "Resolve with copyright verdict (policing badge required)",
        "operationId": "resolveCopyrightReview",
        "security": [{ "bearerAuth": [] }, { "cookieAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["verdict"],
                "properties": {
                  "verdict": { "type": "string", "enum": ["clear", "rewrite_required", "takedown"] },
                  "verdictNotes": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Review resolved with verdict applied" },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/copyright-reviews/tools/verbatim-search": {
      "get": {
        "tags": ["Copyright Reviews"],
        "summary": "Search for verbatim text matches across chunks (policing badge)",
        "operationId": "verbatimSearch",
        "security": [{ "bearerAuth": [] }, { "cookieAuth": [] }],
        "parameters": [{ "name": "text", "in": "query", "required": true, "schema": { "type": "string", "minLength": 30 } }],
        "responses": {
          "200": { "description": "Matching chunks with position info" }
        }
      }
    },
    "/copyright-reviews/tools/check-sources/{chunkId}": {
      "get": {
        "tags": ["Copyright Reviews"],
        "summary": "Check source citation resolvability (policing badge)",
        "operationId": "checkSources",
        "security": [{ "bearerAuth": [] }, { "cookieAuth": [] }],
        "parameters": [{ "name": "chunkId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "Source resolution status for each citation" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key from account registration"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "aingram_token",
        "description": "JWT session cookie (set on login)"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" }
            }
          }
        }
      }
    },
    "responses": {
      "ValidationError": {
        "description": "Validation error",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
      "NotFound": {
        "description": "Resource not found",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "RateLimited": {
        "description": "Too many requests",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      }
    }
  }
}